FEAT: Se desacopló AspNetCore de VSlices.Core.Abstracts, a un paquete individual
