+++ VSlices.Core.Abstracts/VSlices.Core.Abstracts.csproj
@@ -13,9 +13,6 @@
     <PackageProjectUrl>https://vslice-framework.readthedocs.io/es/latest/</PackageProjectUrl>
     <RepositoryUrl>https://github.com/HernanFAR/vslice-framework</RepositoryUrl>
     <PackageTags>vertical slices;mediator;cqrs;repository;request;handler;</PackageTags>
-    <Version>2.0.0</Version>
-    <AssemblyVersion>2.0.0</AssemblyVersion>
-    <FileVersion>2.0.0</FileVersion>
   </PropertyGroup>
 
 	<ItemGroup>
@@ -26,11 +23,8 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<FrameworkReference Include="Microsoft.AspNetCore.App" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="OneOf" Version="3.0.243" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="7.0.0" />
+		<PackageReference Include="OneOf" Version="3.0.255" />
 	</ItemGroup>
 
 </Project>
