+++ VSlices.Core.DataAccess.EntityFramework/VSlices.Core.DataAccess.EntityFramework.csproj
@@ -23,7 +23,7 @@
 	</ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.5" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.8" />
   </ItemGroup>
 
   <ItemGroup>
