+++ VSlices.Core.Presentation.AspNetCore/VSlices.Core.Presentation.AspNetCore.csproj
@@ -1,0 +1,26 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+	  <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+	  <Authors>Enryu20</Authors>
+	  <Product>VSlices</Product>
+	  <PackageReadmeFile>README.md</PackageReadmeFile>
+	  <Description>Framework web de alto nivel C#, fuertemente inspirado en la arquitectura "Vertical Slices", MediatR y la programación funcional</Description>
+	  <Copyright>MIT</Copyright>
+	  <PackageProjectUrl>https://vslice-framework.readthedocs.io/es/latest/</PackageProjectUrl>
+	  <RepositoryUrl>https://github.com/HernanFAR/vslice-framework</RepositoryUrl>
+	  <PackageTags>vertical slices;mediator;cqrs;repository;request;handler;</PackageTags>
+  </PropertyGroup>
+
+	<ItemGroup>
+		<FrameworkReference Include="Microsoft.AspNetCore.App" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\VSlices.Core.Abstracts\VSlices.Core.Abstracts.csproj" />
+	</ItemGroup>
+
+</Project>
