+++ VSlices.sln
@@ -47,6 +47,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "BusinessLogic", "BusinessLo
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "BusinessLogic", "BusinessLogic", "{C9B0D647-20C1-4D71-997D-5DE29813E28A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VSlices.Core.Presentation.AspNetCore", "VSlices.Core.Presentation.AspNetCore\VSlices.Core.Presentation.AspNetCore.csproj", "{4C852170-F4F2-4C8D-888E-EA2090E752F5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VSlices.Core.Presentation.AspNetCore.UnitTests", "VSlices.Core.Presentation.AspNetCore.UnitTests\VSlices.Core.Presentation.AspNetCore.UnitTests.csproj", "{EA61F805-4E02-426F-997F-E740A511B5DF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -109,6 +113,14 @@ Global
 		{4A044F04-ACCB-463E-8033-009FE1E1BAFC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4A044F04-ACCB-463E-8033-009FE1E1BAFC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4A044F04-ACCB-463E-8033-009FE1E1BAFC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4C852170-F4F2-4C8D-888E-EA2090E752F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4C852170-F4F2-4C8D-888E-EA2090E752F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4C852170-F4F2-4C8D-888E-EA2090E752F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4C852170-F4F2-4C8D-888E-EA2090E752F5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EA61F805-4E02-426F-997F-E740A511B5DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EA61F805-4E02-426F-997F-E740A511B5DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EA61F805-4E02-426F-997F-E740A511B5DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EA61F805-4E02-426F-997F-E740A511B5DF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -134,6 +146,8 @@ Global
 		{671D1A3F-CE83-4CA1-90CD-F42C1F124E28} = {6F484706-EDFC-49EF-A419-D3F6943C2E36}
 		{A204570A-C4BD-4F8D-9368-A6618A7F3A7F} = {0D8209DB-7023-48E6-85ED-028BF82BA802}
 		{C9B0D647-20C1-4D71-997D-5DE29813E28A} = {45B6CAFB-D06E-4530-8546-F9C7BAE95D96}
+		{4C852170-F4F2-4C8D-888E-EA2090E752F5} = {0D8209DB-7023-48E6-85ED-028BF82BA802}
+		{EA61F805-4E02-426F-997F-E740A511B5DF} = {45B6CAFB-D06E-4530-8546-F9C7BAE95D96}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {653AE7A1-A4C4-42F2-9804-FE7E032C3E8C}
